pyproject.toml
[project]
name = "llm-p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt<4",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.15.11",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.45.0",
]
